Change mechanism for calling delegates to be more thread safe
The previous mechanism can end up calling a delegate that has been removed by a setDelegate:nil, and hence may already be deallocated.

Now we run a callback on the mainthread, which only dereferences the delegate/queue and selectors to be called when it runs on the main thread.

The authenticationdelegate code is still using the old method. This should be looked at separately; it is slightly more complex as it needs to know if a delegate exists to be called.